+++ framework/core/subsystems/expTheme.php
@@ -208,7 +208,7 @@ class expTheme {
 
    		$validateTheme['footerinfo'] = true;
 
-   		if (!empty($user->getsToolbar) && PRINTER_FRIENDLY != 1 && EXPORT_AS_PDF != 1 && !defined('SOURCE_SELECTOR') && empty($params['hide-slingbar'])) {
+   		if (!empty($user->getsToolbar) && !$user->globalPerm('hide_slingbar') && PRINTER_FRIENDLY != 1 && EXPORT_AS_PDF != 1 && !defined('SOURCE_SELECTOR') && empty($params['hide-slingbar'])) {
    			self::module(array("controller"=>"administration","action"=>"toolbar","source"=>"admin"));
    		}